lab/bodega/requirements.txt
boto3==1.7.4
celery[tblib]==4.0.2
Django==1.10.4
django-braces==1.11.0
django_celery_beat==1.0.1
django-filter==1.0.1
django-memoize==2.1.0
django-oauth-toolkit==0.12.0
django-permission==0.9.3
django-pyodbc-azure==1.10.4.0
django-rest-framework-social-oauth2==1.0.5
djangorestframework==3.5.3
future==0.16.0
jenkinsapi==0.3.3
kombu==4.1.0
kubernetes==1.0.0
oauth2client==4.1.0
paramiko==1.15.1
pycrypto==2.6.1
pyodbc==4.0.21
pytz==2016.10
pyvmomi==6.5.0.2017.5-1
pyyaml==3.12
rest-framework-generic-relations==1.1.0
retrying==1.3.3
social-auth-app-django==0.1.0
social-auth-core==0.2.1
slacker==0.9.30
statsd==3.2.1
# wget was added to support RkLabBaseJenkins
wget==3.2
